Fix the EEPROM layout, so we remain compatible
We need to allocate space for the IdleLEDs setting within the scope of the
`onSetup()` chain, we can't do it after `Kaleidoscope.setup()`, because
`Kaleidoscope.setup()` will also call `device().setup()`, which in this case
will allocate two bytes for the keyscanner settings, thus effectively swapping
that setting with IdleLEDs'.

To combat this, we will simply use the `PersistentIdleLEDs` plugin at the
appropriate position, to maintain EEPROM layout compatibility. The
`RaiseIdleLEDs` plugin, which we put at the end will proxy all key events to
`PersistentIdleLEDs`. This way the IdleLEDs plugin can still see all key events,
but allocation will happen at the right place.

The downside is that IdleLEDs will now act twice on most key presses, but
thankfully that has no ill side effects, and is only a little more costly than a
no-op.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>